items: extract get_authorization_level from get_authorized_items
to allow independant use
