+++ server/controllers/items/lib/get_authorized_items.js
@@ -1,26 +1,11 @@
-const __ = require('config').universalPath
-const getByAccessLevel = require('./get_by_access_level')
-const { areFriendsOrGroupCoMembers } = __.require('controllers', 'user/lib/relations_status')
-const groups_ = __.require('controllers', 'groups/lib/groups')
+const getByAuthorizationLevel = require('./get_by_authorization_level')
+const { byUser, byGroup } = require('./get_authorization_level')
 
 // Return what the reqUserId user is allowed to see from a user or a group inventory
 module.exports = {
-  byUser: (userId, reqUserId) => {
-    if (userId === reqUserId) return getByAccessLevel.private(userId)
-    if (!reqUserId) return getByAccessLevel.public(userId)
+  byUser: (userId, reqUserId) => byUser(userId, reqUserId).then(getItems),
 
-    return areFriendsOrGroupCoMembers(userId, reqUserId)
-    .then(usersAreFriendsOrGroupCoMembers => {
-      if (usersAreFriendsOrGroupCoMembers) return getByAccessLevel.network(userId)
-      else return getByAccessLevel.public(userId)
-    })
-  },
-
-  byGroup: (groupId, reqUserId) => {
-    return groups_.getGroupMembersIds(groupId)
-    .then(allGroupMembers => {
-      if (reqUserId && allGroupMembers.includes(reqUserId)) return getByAccessLevel.network(allGroupMembers)
-      else return getByAccessLevel.public(allGroupMembers)
-    })
-  }
+  byGroup: (groupId, reqUserId) => byGroup(groupId, reqUserId).then(getItems)
 }
+
+const getItems = ({ authorizationLevel, usersIds }) => getByAuthorizationLevel[authorizationLevel](usersIds)
